Add: menu.js용 라우터 설정
