+++ src/components/menu/addMenu.js
@@ -1,0 +1,22 @@
+//src/components/Play.js
+
+import axios from "axios";
+import { useEffect, useRef, useState } from "react";
+import { Button, Container, Form } from "react-bootstrap";
+
+
+
+
+function AddMenu(){
+ 
+
+    return(
+     <div>
+        
+     </div>
+       
+    )
+}
+
+export default AddMenu
+
